requirements.txt
python-dotenv~=1.1.0
selenium~=4.33.0
webdriver-manager~=4.0.2
requests
beautifulsoup4
pandas